bakend/requirements.txt
aioredis
asgiref
async-timeout
attrs
autobahn
Automat
cachetools
certifi
cffi
channels
channels-redis
charset-normalizer
constantly
cryptography
daphne
dj-database-url
Django
django-appconf
django-cors-headers
django-storages
djangorestframework
gevent
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
greenlet
gunicorn
hiredis
httplib2
hyperlink
idna
incremental
msgpack
oauth2client
packaging
Pillow
protobuf
psycopg2
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
pyOpenSSL
pyparsing
python-dateutil
pytz
requests
rsa
service-identity
six
sqlparse==0.4.1
timeago
Twisted
twisted-iocpsupport
txaio
typing-extensions
uritemplate
urllib3
zope.event
zope.interface
